[#5] Added Editor component
